// convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * Generated by convex@1.2.1.
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";
import type * as _helpers_ai_helper from "../_helpers/ai_helper";
import type * as _helpers_helpers from "../_helpers/helpers";
import type * as _helpers_statusCodes from "../_helpers/statusCodes";
import type * as actions from "../actions";
import type * as ai from "../ai";
import type * as chat from "../chat";
import type * as game from "../game";
import type * as host from "../host";
import type * as monuments from "../monuments";
import type * as player from "../player";

/**
 * A utility for referencing Convex functions in your app's API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
declare const fullApi: ApiFromModules<{
  "_helpers/ai_helper": typeof _helpers_ai_helper;
  "_helpers/helpers": typeof _helpers_helpers;
  "_helpers/statusCodes": typeof _helpers_statusCodes;
  actions: typeof actions;
  ai: typeof ai;
  chat: typeof chat;
  game: typeof game;
  host: typeof host;
  monuments: typeof monuments;
  player: typeof player;
}>;
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;